# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de Alvaro's Laravel Api",
        "description": "Esta es una api simple hecha con laravel 8.0 con autenticación realizada con Oauth 2.0",
        "contact": {
            "email": "alvarogarcia1010@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Alvaro's Api"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Lista de productos",
                "description": "Obtengo una lista de productos con paginación",
                "operationId": "getArticles",
                "parameters": [
                    {
                        "name": "page[size]",
                        "in": "path",
                        "description": "Tamaño de la pagina",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page[number]",
                        "in": "path",
                        "description": "Número de pagina",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "path",
                        "description": "Buscador de productos",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "path",
                        "description": "Ordernar por campo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acceso no autorizado, para acceder a este recurso debe iniciar sesión"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualización de productos",
                "description": "Ruta para actualizar un producto",
                "operationId": "updateArticle",
                "requestBody": {
                    "description": "Información a actualizar del producto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "quantity",
                                    "price"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Producto 1"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "price": {
                                        "type": "float",
                                        "example": "35.50"
                                    },
                                    "remark": {
                                        "type": "string",
                                        "example": "Esta es una descripción"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "URL"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acceso no autorizado, para acceder a este recurso debe iniciar sesión"
                    },
                    "422": {
                        "description": "Campos no validos o faltan campos requeridos"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Creación de productos",
                "description": "Ruta para crear un nuevo producto",
                "operationId": "createArticle",
                "requestBody": {
                    "description": "Información del nuevo producto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "quantity",
                                    "price"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Producto 1"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "price": {
                                        "type": "float",
                                        "example": "35.50"
                                    },
                                    "remark": {
                                        "type": "string",
                                        "example": "Esta es una descripción"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "URL"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acceso no autorizado, para acceder a este recurso debe iniciar sesión"
                    },
                    "422": {
                        "description": "Campos no validos o faltan campos requeridos"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/articles/{articleId}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener producto por id",
                "description": "Esta ruta obtiene un producto por el id",
                "operationId": "getArticle",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "ID del producto a retornar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acceso no autorizado, para acceder a este recurso debe iniciar sesión"
                    },
                    "404": {
                        "description": "Artículo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "description": "Esta ruta elimina el producto con el id pasado por parametro",
                "operationId": "deleteArticle",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID del prodcuto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acceso no autorizado, para acceder a este recurso debe iniciar sesión"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}